housing_pricing/requirements.txt
scikit-learn
pandas
seaborn
flask 
flask-cors
matplotlib
Jinja2==3.1.2
PyYAML
python-box==6.0.2
ensure


-e .